Align renderer and app themes, adjust text field borders to be darker
