+++ apps/smart-forms-app/package.json
@@ -36,9 +36,9 @@
     "@iconify/react": "^5.2.0",
     "@microlink/react-json-view": "^1.26.1",
     "@monaco-editor/react": "^4.7.0",
-    "@mui/icons-material": "^7.0.1",
-    "@mui/lab": "^7.0.0-beta.10",
-    "@mui/material": "^7.0.1",
+    "@mui/icons-material": "^7.1.1",
+    "@mui/lab": "^7.0.0-beta.13",
+    "@mui/material": "^7.1.1",
     "@sentry/react": "^9.3.0",
     "@sentry/tracing": "^7.120.3",
     "@tanstack/react-query": "^5.72.1",
