+++ apps/smart-forms-app/src/theme/overrides/combinedComponentOverrides.ts
@@ -15,16 +15,18 @@
  * limitations under the License.
  */
 
-import type { Theme } from '@mui/material';
+import type { Theme } from '@mui/material/styles';
+import { backdropOverride } from './backdropOverride.ts';
+import { dialogOverride } from './dialogOverride.ts';
+import { rendererThemeComponentOverrides } from '@aehrc/smart-forms-renderer';
 
-export default function Autocomplete(theme: Theme) {
+export function combinedComponentOverrides(theme: Theme) {
+  // @ts-ignore baffling TS error for the type Theme - could be due to using different MUI versions, last checked both are using v7.1.1 but still giving an error
+  const rendererOverrides = rendererThemeComponentOverrides(theme);
+  const appOverrides = Object.assign(backdropOverride(theme), dialogOverride(theme));
+
   return {
-    MuiAutocomplete: {
-      styleOverrides: {
-        paper: {
-          boxShadow: theme.shadows[12]
-        }
-      }
-    }
+    ...rendererOverrides,
+    ...appOverrides
   };
 }
