+++ packages/smart-forms-renderer/src/theme/index.ts
@@ -1,2 +1,18 @@
+// Export whole RendererThemeProvider
 export { RendererThemeProvider } from './RendererThemeProvider';
-export { themeOptions } from './themeOptions';
+
+// Export whole component overrides
+export { rendererThemeComponentOverrides } from './overrides/rendererThemeComponentOverrides';
+
+// Export individual component overrides
+export { accordionOverride } from './overrides/accordionOverride';
+export { autocompleteOverride } from './overrides/autocompleteOverride';
+export { buttonOverride } from './overrides/buttonOverride';
+export { cardOverride } from './overrides/cardOverride';
+export { inputOverride } from './overrides/inputOverride';
+export { paperOverride } from './overrides/paperOverride';
+export { speedDialOverride } from './overrides/speedDialOverride';
+export { tableOverride } from './overrides/tableOverride';
+
+// Export themeOptions
+export { rendererThemeOptions } from './rendererThemeOptions';
