--- packages/smart-forms-renderer/src/theme/overrides/Accordion.ts
@@ -1,39 +1,0 @@
-/*
- * Copyright 2024 Commonwealth Scientific and Industrial Research
- * Organisation (CSIRO) ABN 41 687 119 230.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-import type { Theme } from '@mui/material';
-import { alpha } from '@mui/material/styles';
-import { grey } from '@mui/material/colors';
-
-export default function Accordion(theme: Theme) {
-  return {
-    MuiAccordion: {
-      styleOverrides: {
-        root: {
-          boxShadow: `0 0 2px 0 ${alpha(grey[500], 0.2)}, 0 12px 24px -4px ${alpha(
-            grey[500],
-            0.16
-          )}`,
-          borderRadius: Number(theme.shape.borderRadius) * 2,
-          '&:before': {
-            display: 'none'
-          }
-        }
-      }
-    }
-  };
-}
