--- packages/smart-forms-renderer/src/theme/overrides/Input.ts
@@ -1,94 +1,0 @@
-/*
- * Copyright 2024 Commonwealth Scientific and Industrial Research
- * Organisation (CSIRO) ABN 41 687 119 230.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-import type { Theme } from '@mui/material/styles';
-import { alpha } from '@mui/material/styles';
-
-export default function Input(theme: Theme) {
-  return {
-    MuiInputBase: {
-      styleOverrides: {
-        root: {
-          '&.Mui-disabled': {
-            '& svg': { color: theme.palette.text.disabled }
-          }
-        },
-        input: {
-          '&::placeholder': {
-            opacity: 1,
-            color: theme.palette.text.disabled
-          }
-        }
-      }
-    },
-    MuiInput: {
-      styleOverrides: {
-        underline: {
-          '&:before': {
-            borderBottomColor: alpha(theme.palette.grey[500], 0.56)
-          }
-        }
-      }
-    },
-    MuiFilledInput: {
-      styleOverrides: {
-        root: {
-          backgroundColor: alpha(theme.palette.grey[500], 0.12),
-          '&:hover': {
-            backgroundColor: alpha(theme.palette.grey[500], 0.16)
-          },
-          '&.Mui-focused': {
-            backgroundColor: theme.palette.action.focus
-          },
-          '&.Mui-disabled': {
-            backgroundColor: theme.palette.action.disabledBackground
-          }
-        },
-        underline: {
-          '&:before': {
-            borderBottomColor: alpha(theme.palette.grey[500], 0.56)
-          }
-        }
-      }
-    },
-    MuiOutlinedInput: {
-      styleOverrides: {
-        root: {
-          '&.Mui-disabled': {
-            '& .MuiOutlinedInput-notchedOutline': {
-              borderColor: theme.palette.action.disabledBackground
-            }
-          },
-          '&.Mui-readOnly': {
-            '& .MuiOutlinedInput-notchedOutline': {
-              borderColor: 'rgb(220, 223, 228)'
-            },
-            '&:hover .MuiOutlinedInput-notchedOutline': {
-              borderColor: 'rgb(220, 223, 228)'
-            },
-            '&.Mui-focused .MuiOutlinedInput-notchedOutline': {
-              borderColor: 'rgb(220, 223, 228)'
-            },
-            '&.Mui-visited .MuiOutlinedInput-notchedOutline': {
-              borderColor: 'rgb(220, 223, 228)'
-            }
-          }
-        }
-      }
-    }
-  };
-}
