+++ packages/smart-forms-renderer/src/theme/overrides/accordionOverride.ts
@@ -15,13 +15,13 @@
  * limitations under the License.
  */
 
-import type { Theme } from '@mui/material';
+import type { Theme } from '@mui/material/styles';
 import { alpha } from '@mui/material/styles';
 import { grey } from '@mui/material/colors';
 
-export default function Card(theme: Theme) {
+export function accordionOverride(theme: Theme) {
   return {
-    MuiCard: {
+    MuiAccordion: {
       styleOverrides: {
         root: {
           boxShadow: `0 0 2px 0 ${alpha(grey[500], 0.2)}, 0 12px 24px -4px ${alpha(
@@ -29,26 +29,9 @@ export default function Card(theme: Theme) {
             0.16
           )}`,
           borderRadius: Number(theme.shape.borderRadius) * 2,
-          position: 'relative',
-          zIndex: 0 // Fix Safari overflow: hidden with border radius
-        }
-      }
-    },
-    MuiCardHeader: {
-      defaultProps: {
-        titleTypographyProps: { variant: 'h6' },
-        subheaderTypographyProps: { variant: 'body2' }
-      },
-      styleOverrides: {
-        root: {
-          padding: theme.spacing(3, 3, 0)
-        }
-      }
-    },
-    MuiCardContent: {
-      styleOverrides: {
-        root: {
-          padding: theme.spacing(3)
+          '&:before': {
+            display: 'none'
+          }
         }
       }
     }
