+++ packages/smart-forms-renderer/src/theme/overrides/cardOverride.ts
@@ -17,36 +17,38 @@
 
 import type { Theme } from '@mui/material/styles';
 import { alpha } from '@mui/material/styles';
+import { grey } from '@mui/material/colors';
 
-export default function Button(theme: Theme) {
+export function cardOverride(theme: Theme) {
   return {
-    MuiButton: {
+    MuiCard: {
       styleOverrides: {
         root: {
-          '&:hover': {
-            boxShadow: 'none'
-          }
-        },
-        sizeLarge: {
-          height: 48
-        },
-        containedInherit: {
-          color: theme.palette.grey[800],
-          boxShadow: theme.shadows[8],
-          '&:hover': {
-            backgroundColor: theme.palette.grey[400]
-          }
-        },
-        outlinedInherit: {
-          border: `1px solid ${alpha(theme.palette.grey[500], 0.32)}`,
-          '&:hover': {
-            backgroundColor: theme.palette.action.hover
-          }
-        },
-        textInherit: {
-          '&:hover': {
-            backgroundColor: theme.palette.action.hover
-          }
+          boxShadow: `0 0 2px 0 ${alpha(grey[500], 0.2)}, 0 12px 24px -4px ${alpha(
+            grey[500],
+            0.16
+          )}`,
+          borderRadius: Number(theme.shape.borderRadius) * 2,
+          position: 'relative',
+          zIndex: 0 // Fix Safari overflow: hidden with border radius
+        }
+      }
+    },
+    MuiCardHeader: {
+      defaultProps: {
+        titleTypographyProps: { variant: 'h6' },
+        subheaderTypographyProps: { variant: 'body2' }
+      },
+      styleOverrides: {
+        root: {
+          padding: theme.spacing(3, 3, 0)
+        }
+      }
+    },
+    MuiCardContent: {
+      styleOverrides: {
+        root: {
+          padding: theme.spacing(3)
         }
       }
     }
