+++ packages/smart-forms-renderer/src/theme/overrides/tableOverride.ts
@@ -15,17 +15,28 @@
  * limitations under the License.
  */
 
-import type { Theme } from '@mui/material';
+import type { Theme } from '@mui/material/styles';
+import { alpha } from '@mui/material/styles';
+import { grey } from '@mui/material/colors';
 
-export default function Typography(theme: Theme) {
+export function tableOverride(theme: Theme) {
   return {
-    MuiTypography: {
+    MuiTableCell: {
       styleOverrides: {
-        paragraph: {
-          marginBottom: theme.spacing(2)
-        },
-        gutterBottom: {
-          marginBottom: theme.spacing(1)
+        head: {
+          color: theme.palette.text.secondary,
+          backgroundColor: '#F4F6F8',
+          borderBottom: '1px solid #e5e7eb'
+        }
+      }
+    },
+    MuiTableContainer: {
+      styleOverrides: {
+        root: {
+          boxShadow: `0 0 2px 0 ${alpha(grey[500], 0.2)}, 0 12px 24px -4px ${alpha(
+            grey[500],
+            0.16
+          )}`
         }
       }
     }
